bitfinex WS sub still disconnects after a while

+++ fetchers/ws/updater.py
@@ -9,6 +9,7 @@ from common.helpers.datetimehelpers import milliseconds_to_datetime
 from common.config.constants import (
     REDIS_HOST, REDIS_PASSWORD, REDIS_DELIMITER, DBCONNECTION, OHLCVS_TABLE
 )
+from common.utils.logutils import create_logger
 from fetchers.config.constants import (
     WS_SUB_LIST_REDIS_KEY, WS_SUB_PREFIX, WS_SUB_PROCESSING_REDIS_KEY
 )
@@ -16,7 +17,7 @@ from fetchers.config.queries import PSQL_INSERT_IGNOREDUP_QUERY
 from fetchers.helpers.dbhelpers import psql_bulk_insert
 
 
-UPDATE_FREQUENCY_SECS = 60
+UPDATE_FREQUENCY_SECS = 30
 
 class OHLCVWebsocketUpdater:
     def __init__(self):
@@ -27,6 +28,8 @@ class OHLCVWebsocketUpdater:
             decode_responses=True
         )
 
+        self.logger = create_logger("updater_websocket")
+
     def update(self):
         '''
         Collects ohlcv data in ws sub Redis keys and inserts them
@@ -36,8 +39,10 @@ class OHLCVWebsocketUpdater:
         psql_conn = psycopg2.connect(DBCONNECTION)
         try:
             while True:
+                self.logger.info("Collecting subscribed OHLCV data in Redis")
                 ohlcvs_table_insert = []
                 # Loop thru all keys in sub list
+                self.logger.info(f"Length of WS sub list: {self.redis_client.scard(WS_SUB_LIST_REDIS_KEY)}")
                 for key in self.redis_client.smembers(WS_SUB_LIST_REDIS_KEY):
                     try:
                         exchange, base_id, quote_id = \
@@ -47,7 +52,7 @@ class OHLCVWebsocketUpdater:
                         #   sort timestamps ascending,  exclude the latest one,
                         #   prepare ohlcv rows to insert
                         #   then delete data related to inserted timestamps
-                        if data and len(data.keys()) > 1:
+                        if data and len(data) > 1:
                             ts_sorted = sorted(data.keys())
                             ts_to_insert = ts_sorted[:-1]
                             for ts in ts_to_insert:
@@ -74,18 +79,22 @@ class OHLCVWebsocketUpdater:
                                 )
                             self.redis_client.hdel(key, *ts_to_insert)
                     except Exception as exc:
-                        print(f"EXCEPTION: {exc}")
+                        self.logger.warning(f"EXCEPTION: {exc}")
                 # Bulk insert
                 try:
-                    psql_bulk_insert(
+                    success = psql_bulk_insert(
                         psql_conn,
                         ohlcvs_table_insert,
                         OHLCVS_TABLE,
                         insert_ignoredup_query = PSQL_INSERT_IGNOREDUP_QUERY
                     )
-                    self.redis_client.delete(WS_SUB_PROCESSING_REDIS_KEY)
+                    # If success, clean up
+                    # TODO: add a process to clean up if NOT success
+                    if success:
+                        self.redis_client.delete(WS_SUB_PROCESSING_REDIS_KEY)
+                        self.logger.info("Updated OHLCV to PSQL db")
                 except Exception as exc:
-                    print(f"EXCEPTION: {exc}")
+                    self.logger.warning(f"EXCEPTION: {exc}")
                 time.sleep(UPDATE_FREQUENCY_SECS)
         finally:
             psql_conn.close()
